fixed: CV2VideoHandler can't handle non-latin symbols in filenames fixed methods moved from utils to handler as static
