+++ sinner/handlers/frame/CV2VideoHandler.py
@@ -1,16 +1,18 @@
 import glob
 import os.path
+import platform
 from pathlib import Path
 from typing import List
 
 import cv2
 from cv2 import VideoCapture
+from numpy import fromfile, uint8
 from tqdm import tqdm
 
 from sinner.Status import Mood
 from sinner.handlers.frame.BaseFrameHandler import BaseFrameHandler
-from sinner.typing import NumeratedFrame, NumeratedFramePath
-from sinner.utilities import write_image, get_file_name
+from sinner.typing import NumeratedFrame, NumeratedFramePath, Frame
+from sinner.utilities import get_file_name
 from sinner.validators.AttributeLoader import Rules
 
 
@@ -87,7 +89,7 @@ class CV2VideoHandler(BaseFrameHandler):
                 ret, frame = capture.read()
                 if not ret:
                     break
-                write_image(frame, os.path.join(path, str(i + 1).zfill(filename_length) + ".png"))
+                self.write_image(frame, os.path.join(path, str(i + 1).zfill(filename_length) + ".png"))
                 progress.update()
                 i += 1
             capture.release()
@@ -111,12 +113,12 @@ class CV2VideoHandler(BaseFrameHandler):
         try:
             Path(os.path.dirname(filename)).mkdir(parents=True, exist_ok=True)
             frame_files = glob.glob(os.path.join(glob.escape(from_dir), '*.png'))
-            first_frame = cv2.imread(frame_files[0])
+            first_frame = self.read_image(frame_files[0])
             height, width, channels = first_frame.shape
             fourcc = self.suggest_codec()
             video_writer = cv2.VideoWriter(filename, fourcc, self.output_fps, (width, height))
             for frame_path in frame_files:
-                frame = cv2.imread(frame_path)
+                frame = self.read_image(frame_path)
                 video_writer.write(frame)
             video_writer.release()
             return True
@@ -132,3 +134,22 @@ class CV2VideoHandler(BaseFrameHandler):
                 self.update_status(message=f"Suggested codec: {fourcc}", mood=Mood.NEUTRAL)
                 return fourcc
         raise NotImplementedError('No supported codecs found')
+
+    @staticmethod
+    def read_image(path: str) -> Frame:
+        if platform.system().lower() == 'windows':  # issue #511
+            image = cv2.imdecode(fromfile(path, dtype=uint8), cv2.IMREAD_UNCHANGED)
+            if image.shape[2] == 4:  # fixes the alpha-channel issue
+                image = cv2.cvtColor(image, cv2.COLOR_BGRA2BGR)
+            return image
+        else:
+            return cv2.imread(path)
+
+    @staticmethod
+    def write_image(image: Frame, path: str) -> bool:
+        if platform.system().lower() == 'windows':  # issue #511
+            is_success, im_buf_arr = cv2.imencode(".png", image)
+            im_buf_arr.tofile(path)
+            return is_success
+        else:
+            return cv2.imwrite(path, image)
